Add files via upload
Welcome to my GitHub repository dedicated to my diabetes detection project in women. Leveraging the power of machine learning, this project aims to build an accurate predictive model for assessing the likelihood of diabetes in female individuals based on a diverse set of health-related features. From data collection and preprocessing to model selection and deployment, this repository encompasses the entire workflow. Our meticulously curated dataset, carefully chosen machine learning algorithms, and a commitment to ethical considerations come together to offer a comprehensive solution for healthcare professionals and researchers interested in diabetes risk assessment for women. I invite you to explore the code, the model's performance, and the associated documentation to gain insights into this crucial healthcare application.